+++ src/arche/tools/schema.py
@@ -1,13 +1,19 @@
+from collections import defaultdict
 from dataclasses import dataclass
 import json
 import pprint
 import random
-from typing import Any, Dict, List, Optional
+from typing import Any, Deque, Dict, List, Optional
 
+from arche.readers.items import RawItems
 from arche.readers.schema import Schema
 from arche.schema_definitions import extension
 from arche.tools import api, helpers
+import fastjsonschema
 from genson import SchemaBuilder
+from jsonschema import FormatChecker, validators
+import pandas as pd
+from tqdm import tqdm_notebook
 
 
 @dataclass
@@ -57,6 +63,8 @@ def create_json_schema(
     samples = []
     for n in item_numbers:
         item = api.get_items(source_key, start_index=n, count=1, p_bar=None)[0]
+        item.pop("_type", None)
+        item.pop("_key", None)
         samples.append(item)
 
     return infer_schema(samples)
@@ -80,3 +88,89 @@ def set_item_no(items_count: int) -> List[int]:
     if items_count <= 4:
         return [i for i in range(items_count)]
     return random.sample(range(0, items_count), 4)
+
+
+def fast_validate(
+    schema: Schema, raw_items: RawItems, keys: pd.Index
+) -> Dict[str, set]:
+    """Verify items one by one. It stops after the first error in an item in most cases.
+    Faster than jsonschema validation
+
+    Args:
+        schema: a JSON schema
+        raw_items: a raw data to validate one by one
+        keys: keys corresponding to raw_items index
+
+    Returns:
+        A dictionary of errors with message and item keys
+    """
+    errors = defaultdict(set)
+
+    validate = fastjsonschema.compile(schema)
+    for i, raw_item in enumerate(
+        tqdm_notebook(raw_items, desc="Fast Schema Validation")
+    ):
+        raw_item.pop("_type", None)
+        raw_item.pop("_key", None)
+        try:
+            validate(raw_item)
+        except fastjsonschema.JsonSchemaException as error:
+            errors[str(error)].add(keys[i])
+    return errors
+
+
+def full_validate(
+    schema: Schema, raw_items: RawItems, keys: pd.Index
+) -> Dict[str, set]:
+    """This function uses jsonschema validator which returns all found error per item.
+    See `fast_validate()` for arguments descriptions.
+    """
+    errors = defaultdict(set)
+
+    validator = validators.validator_for(schema)(schema)
+    validator.format_checker = FormatChecker()
+    for i, raw_item in enumerate(
+        tqdm_notebook(raw_items, desc="JSON Schema Validation")
+    ):
+        raw_item.pop("_type", None)
+        raw_item.pop("_key", None)
+        for e in validator.iter_errors(raw_item):
+            error = format_validation_message(
+                e.message, e.path, e.schema_path, e.validator
+            )
+            errors[error].add(keys[i])
+    return errors
+
+
+def format_validation_message(
+    error_msg: str, path: Deque, schema_path: Deque, validator: str
+) -> str:
+    str_path = "/".join(p for p in path if isinstance(p, str))
+    schema_path = "/".join(p for p in schema_path)
+
+    if validator == "anyOf":
+        if str_path:
+            return f"'{str_path}' does not satisfy 'schema/{schema_path}'"
+        else:
+            return f"'schema/{schema_path}' failed"
+
+    if "Additional properties are not allowed" in error_msg:
+        return error_msg[: error_msg.index(" (")]
+
+    if not str_path:
+        return error_msg
+
+    for path_message in [
+        "is not of type",
+        "does not match",
+        "is not one of",
+        "is not a",
+        "is greater than the maximum of",
+        "is less than the minimum of",
+        "is too long",
+        "is too short",
+    ]:
+        if path_message in error_msg:
+            return f"{str_path} {error_msg[error_msg.index(path_message) :]}"
+
+    return f"{str_path} - {error_msg}"