+++ CHANGES.md
@@ -16,6 +16,8 @@ Note that the top-most release is changes in the unreleased master branch on Git
 ### Changed
 - `Arche.report_all()` does not shorten report by default, added `short` parameter.
 - `expand=True` which enables nested data flattening is more than 100x faster and consumes ~2x less memory than before, #94
+- Data is consistent with Dash and Spidermon: `_type, _key` fields are dropped from dataframe, raw data, basic schema, #104, #106
+- `df.index` now stores `_key` instead
 - `basic_json_schema()` works with `deleted` jobs
 ### Fixed
 - `Arche.glance()`, #88